[temporal] Remove custom time zones and calendars
Normative change from June 2024 TC39. Removes the Temporal.Calendar and
Temporal.TimeZone objects and replaces them with string identifiers. The
calendar and time zone protocols are removed, along with all calls into
user code that were necessary for implementing them.

Internally, calendars are now represented by an integer index. Time
zones are a bit more complicated, but the complication is encapsulated
in TimeZoneDataRecord. A time zone is either a UTC offset (for nameless
time zones such as +02:00) or an integer index for named time zones.

Also removes
- .calendar properties, replaced by .calendarId (except for
  PlainTime.p.calendar which is just removed)
- .getISOFields methods
- ZonedDateTime.p.timeZone, replaced by .timeZoneId

Note this is based on a rough cut that I did, as part of an
investigation into how much space in the V8 binaries was taken up by
calendars and time zones. As such, some operations may deviate from the
current spec text. My understanding is that these will be ripped out
anyway and replaced with temporal_rs, so that is perhaps not
objectionable.

I have verified that all Temporal test262 tests which previously passed,
still pass, and additionally there are about 500 or so previously
failing tests which now pass.